Add component types and break responsabillities of the file
